chore: remove duplicate of assign_table
