+++ vm-circuit/src/chips/execution_chip/lookup_tables/pow2_fixed_table.rs
@@ -1,4 +1,4 @@
-use crate::chips::execution_chip::instructions::common::assign_table;
+use crate::chips::execution_chip::lookup_tables::utils::assign_table;
 use halo2_proofs::arithmetic::FieldExt;
 use halo2_proofs::circuit::Layouter;
 use halo2_proofs::plonk::{ConstraintSystem, Error, Expression, TableColumn};
