+++ vm-circuit/src/chips/execution_chip/lookup_tables/utils.rs
@@ -1,0 +1,35 @@
+use halo2_proofs::arithmetic::FieldExt;
+use halo2_proofs::circuit::{Layouter, Value};
+use halo2_proofs::plonk::{Error, TableColumn};
+
+pub(crate) fn assign_table<F: FieldExt>(
+    layouter: &mut impl Layouter<F>,
+    table_columns: Vec<TableColumn>,
+    values: &Vec<Vec<F>>,
+    table_name: &str,
+) -> Result<(), Error> {
+    for (column_idx, column) in table_columns.into_iter().enumerate() {
+        layouter.assign_table(
+            || format!("{:?}[{}]", table_name, column_idx),
+            |mut table_column| {
+                table_column.assign_cell(
+                    || format!("{:?}[{}][0]", table_name, column_idx),
+                    column,
+                    0,
+                    || Value::known(F::zero()),
+                )?;
+                (0..values.len())
+                    .map(|i| {
+                        table_column.assign_cell(
+                            || format!("{:?}[{}][{}]", table_name, column_idx, i + 1),
+                            column,
+                            i + 1,
+                            || Value::known(values[i][column_idx]),
+                        )
+                    })
+                    .fold(Ok(()), |acc, res| acc.and(res))
+            },
+        )?;
+    }
+    Ok(())
+}
